pyproject.toml
[project]
name = "dfm-sdg"
version = "0.1.0"
description = "Thin commons plus method packs for synthetic data generation."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
sdg = "sdg.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["B", "E4", "E7", "E9", "F", "I", "PIE", "RUF", "UP"]

[tool.setuptools.packages.find]
include = ["sdg*"]
